fix: Double was wrong in three ways at once
Found by `ghc_differential.py` on its first run, through a fixture whose
`expected.txt` had been written to match BHC rather than Haskell.

  * `/` did INTEGER division. `/` and `div` both lowered to `PrimOp::Div`,
    so by codegen the difference was gone and the choice fell to the
    operand TYPES — and in `1 / 3 :: Double` those are integer literals.
    It emitted `sdiv`, so the program printed `0`, agreed that
    `x > 0.3` was False and that `x * 3` was `0.0`. `1.0 / 3.0` was
    right, which is what hid it. `/` now has its own `FDiv`, which is
    fractional division however the operands look.

  * Every floating literal was truncated to SINGLE precision. Core's
    `Literal::Float` was `f32` and hir-to-core wrote each literal through
    `as f32`, so `0.001 :: Double` reached codegen as the nearest single
    and printed `1.0000000474974513e-3`. Widened to `f64`; a bare
    floating literal defaults to Double in Haskell, so the wider type is
    the one to carry.

  * `show` truncated to six fractional digits and never used scientific
    notation: `1.414214` for `sqrt 2`, `0.001` for `1.0e-3`. It now
    follows Haskell — shortest round-trip, fixed in [0.1, 10^7),
    scientific outside. The old comment defended six digits as "matching
    the WASM backend and the conformance fixtures", but the fixtures were
    written to match this function: the two agreed with each other and
    neither agreed with Haskell.

Also here:

  * `coerce_value` gained Int -> Float. Making `/` fractional sends an
    integer operand into a double slot where it used to stay on the
    integer path, which broke `Text.Pandoc.ImageSize`.
  * Two `expected.txt` corrected to GHC's output: `float_math` (the
    truncated sqrt) and `wordcount` (`lines "a\nb\n"` is 2, not 3 — a
    fixture nothing asserts, quietly carrying a falsehood).
  * `test_tier2_float_math_wasm` is ignored with its reason: wasm's
    `double_to_str` is hand-emitted WASM and still truncates. The gap is
    recorded rather than hidden behind a wrong expectation.
  * `LlvmContext::new` registers LLVM's targets itself. They were only
    registered by `LlvmBackend::new`, so `test_module_verification`
    passed or failed on whether the binary happened to run another test
    first. It was never a flake; it was order dependence.

Differential: 214 agree, 0 diverge (was 213/1) with no contradicted
expectations. Gated: pandoc 221/221, ladder 31, battery 10, tests 2817/0.

+++ crates/bhc-core/src/lib.rs
@@ -343,8 +343,14 @@ pub enum Literal {
     Int(i64),
     /// Arbitrary precision integer.
     Integer(i128),
-    /// Single-precision float.
-    Float(f32),
+    /// Floating literal, at DOUBLE precision.
+    ///
+    /// This was `f32`, and `hir-to-core` wrote every floating literal through
+    /// `as f32`, so `0.001 :: Double` reached codegen as the nearest single —
+    /// printing `1.0000000474974513e-3`. Haskell's defaulting makes a bare
+    /// floating literal a `Double`, so the wider type is the correct one to
+    /// carry; a `Float` context narrows at use.
+    Float(f64),
     /// Double-precision float.
     Double(f64),
     /// A character.
